modifs sur les tests et certains fichiers sont notés comme non testables

+++ src/generator/GeneratorPreModulesPatchDBController.ts
@@ -1,3 +1,4 @@
+/* istanbul ignore next: not a usefull test to write */
 import { IDatabase } from 'pg-promise';
 
 export default class GeneratorPreModulesPatchDBController {
@@ -15,6 +16,9 @@ export default class GeneratorPreModulesPatchDBController {
 
     public async drop_table(db: IDatabase<any>, table_name: string) {
         try {
+            if ((!table_name) || (table_name.indexOf(';') >= 0) || (table_name.indexOf(' ') >= 0)) {
+                return null;
+            }
             await db.none("DROP TABLE " + table_name + ";");
         } catch (error) {
             console.error('Si erreur != table ou colonne inexistante, il faut regarder manuellement :' + error);
@@ -23,6 +27,12 @@ export default class GeneratorPreModulesPatchDBController {
 
     public async drop_column(db: IDatabase<any>, table_name: string, column_name: string) {
         try {
+            if ((!table_name) || (table_name.indexOf(';') >= 0) || (table_name.indexOf(' ') >= 0)) {
+                return null;
+            }
+            if ((!column_name) || (column_name.indexOf(';') >= 0) || (column_name.indexOf(' ') >= 0)) {
+                return null;
+            }
             await db.none("ALTER TABLE " + table_name + " DROP COLUMN " + column_name + ";");
         } catch (error) {
             console.error('Si erreur != table ou colonne inexistante, il faut regarder manuellement :' + error);
